+++ packaging.R
@@ -16,7 +16,7 @@ use_description(
   fields = list(
     Title = "Kernel SHAP",
     Version = "0.2.0.900",
-    Description = "Multidimensional version of the iterative Kernel SHAP algorithm described in
+    Description = "Multidimensional variant of the iterative Kernel SHAP algorithm described in
     Ian Covert and Su-In Lee (2021) <http://proceedings.mlr.press/v130/covert21a>. 
     SHAP values are calculated iteratively until convergence, along with approximate standard errors. 
     The package allows to work with any model that provides numeric predictions of dimension one or higher.